chore(api): rename CertificationChallengeRepository to ChallengeCalibrationRepository
